22.06.17

+++ CapaEntidad/Campaña.cs
@@ -9,19 +9,28 @@ namespace CapaEntidad
     /// Clase de datos para campañas.
     /// </summary>
 	public class Campaña
-	{
+    {
+        /// <summary>
+        /// Clave foránea del elemento cartelería. Funciona como clave principal de la campaña.
+        /// </summary>
 		[Key, ForeignKey("ElementoCarteleria")]
-		public int refElementoCarteleriaId { get; set; }
+		public Int64 refElementoCarteleriaId { get; set; }
 		public virtual ElementoCarteleria ElementoCarteleria { get; set; }
 
+        /// <summary>
+        /// Tiempo que durará cada imágen hasta la siguiente.
+        /// </summary>
         [Required]
 		public virtual TimeSpan TiempoXImagen { get; set; }
 
+        /// <summary>
+        /// Colección de imagenes a mostrar en la campaña.
+        /// </summary>
         [Required]
 		public virtual ICollection<ImagenCampaña> ListaImagenes { get; set; }
 
         /// <summary>
-        /// Inicializa una instancia de la clase <see cref="CapaEntidad.Campaña"/>.
+        /// Inicializa una instancia de la clase <see cref="Campaña"/>.
         /// </summary>
 		public Campaña()
 		{
@@ -37,6 +46,9 @@ namespace CapaEntidad
 			this.ListaImagenes.Add (new ImagenCampaña (pRuta));
 		}
 
+        /// <summary>
+        /// Representación string de la clase.
+        /// </summary>
         public override string ToString()
         {
             return this.ListaImagenes.Count + " imágenes. " + this.TiempoXImagen.ToString() + " por imágen.";
